Fix workflows (#58)
# Description

Fixing the prefix change made some PRs before. For the workflows, removing the option to deploy the bot to dev until is pushed to develop branch (before is tried to deploy every time that is a PR), and allowing to run the build workflow (run bot and test) for any pull request created.

# Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] Documentation update

# How Has This Been Tested?

Locally tested

# Checklist:

- [ ] I have commented my code, particularly in hard-to-understand areas
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my feature works