Add leetcode gql schemas (#82)
# Description

Adding schemas to be used with [pydantic](https://docs.pydantic.dev/latest/) during the next PR, this will give structure to the graphql response received. 
Also updated discord.py version to 2.4.0 and fixed a test because of that update.

# Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] Documentation update

# How Has This Been Tested?

Not required, to be tested during the next PR when used.

# Checklist:

- [x] I have commented my code, particularly in hard-to-understand areas
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my feature works